+++ src/main/java/com/diviso/purchase/domain/QuotationLine.java
@@ -26,7 +26,7 @@ public class QuotationLine implements Serializable {
     private Integer productPrice;
 
     @Column(name = "product_tax")
-    private Integer productTax;
+    private Float productTax;
 
     @Column(name = "available_quantity")
     private Integer availableQuantity;
@@ -72,16 +72,16 @@ public class QuotationLine implements Serializable {
         this.productPrice = productPrice;
     }
 
-    public Integer getProductTax() {
+    public Float getProductTax() {
         return productTax;
     }
 
-    public QuotationLine productTax(Integer productTax) {
+    public QuotationLine productTax(Float productTax) {
         this.productTax = productTax;
         return this;
     }
 
-    public void setProductTax(Integer productTax) {
+    public void setProductTax(Float productTax) {
         this.productTax = productTax;
     }
